requirements.txt
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.14
django-environ==0.8.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
prompt-toolkit==3.0.29
psycopg2-binary==2.9.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
ujson==5.3.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5